requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
jinja2>=3.1.4
aiofiles>=24.1.0
websockets>=13.0

# Database
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
alembic>=1.14.0

# Auth
pyjwt>=2.9.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# ML - Detection
ultralytics>=8.3.0
opencv-python-headless>=4.8.0

# ML - Identification / Training
torch>=2.1.0
torchvision>=0.16.0
albumentations>=1.4.0
scikit-learn>=1.5.0

# Configuration
pydantic-settings>=2.6.0

# Utilities
numpy>=1.24.0,<2
pillow>=10.0.0
httpx>=0.27.0
